render channel messages as markdown
